Fix an issue caused by Puppet parse order changes
Apparently the parse order has changed when using Hiera. Previously there was no
problem with depending on classes included in Hiera on the lowest level (e.g.
common.yaml). Now, when this class, for example, is included in node's own yaml,
it can no longer find the "localbackups" class, even though the class is clearly
present in the final class list. To fix (or workaround) this commit drops the
localbackups dependency altogether and just ensures that the backup directory is
present.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>